+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `Replicate`, short-hand composition of `take` and `repeat`
 - `Distinct` to produce an `Iterable` of distinct values in another `Iterable`
 - `Sort` and `SortBy` for eagerly, monolithically sorting `Iterable`s and producing `List`s  
+- `IterableLens`, general lenses over `Iterable`   
 
 ## [1.6.2] - 2017-08-20
 ### Fixed